Converts default PushType app to reusable Littlelines CMS
This commit makes numerous adjustments to tailor the PushType CMS to fit
the reusable needs for Littlelines. It adds config options including a
customer logo. It also adds the concept of Inquiries to allow storing
contact form submissions and accessing them via the admin tool. It
allows contact form fields with the name file to be previewed as an
image. This preview is linked to a new Asset show page that gives the
user an option to download the file.